fixed log leak

+++ psd/download_file.php
@@ -9,9 +9,16 @@ set_time_limit(0);
 //$output2 = exec("ls -la", $output, $return);
 
 $id = substr($link, strrpos($link, '/') + 1);
+
+
+
 exec("mkdir downloads; touch downloads/$id;");
 
-$output2 = exec("mkdir temp; cd temp; wget $link;", $output, $return);
+exec("ls temp/$id;", $output, $return);
+
+if ($return == 0) goto end;
+
+$output2 = exec("mkdir temp; cd temp; wget $link 1>/dev/null;", $output, $return);
 
 exec("rm downloads/$id;");
 
@@ -19,5 +26,7 @@ if ($return == 0){
 	exec("mkdir $folder; mv temp/$id $folder/");
 }
 
+
+end:
 
 ?>
